vm: make a virtio-rng device aviable to the vm
Without a good source of entropy the getrandom() syscal will block until the cnrg init is done
which could take very long (in extreme cases over an hour. As in debian buster sshd relies
on getrandom() via openssl 1.1.1, it fails to start before the random init is finished,
so the user is unable to use the VM for a long time.